♻️ refactor: extract call_llm tooling resolution
